Added test cases for ShoppingCart, CartItem entities (repository, service, controller)
